Cloud_Functions/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-auth-oauthlib==1.*
google-api-python-client==2.*
python-dotenv==1.*